Upload `langsmith-walkthrough.ipynb`
Description:
LangChain makes it easy to prototype LLM applications and Agents. However, delivering LLM applications to production can be deceptively difficult. You will likely have to heavily customize and iterate on your prompts, chains, and other components to create a high-quality product.

To aid in this process, we’ve launched LangSmith, a unified platform for debugging, testing, and monitoring your LLM applications.

When might this come in handy? You may find it useful when you want to:

Quickly debug a new chain, agent, or set of tools
Visualize how components (chains, llms, retrievers, etc.) relate and are used
Evaluate different prompts and LLMs for a single component
Run a given chain several times over a dataset to ensure it consistently meets a quality bar
Capture usage traces and using LLMs or analytics pipelines to generate insights

Based on the original:
- https://python.langchain.com/docs/langsmith/walkthrough

Signed-off-by: Daemon <109057945+Daethyra@users.noreply.github.com>